Added completed todo functionality
