+++ src/routes/adminRoutes.ts
@@ -1,5 +1,5 @@
 import express from 'express'
-import { getUsers, listAllTodos, getTodosPerUser, averageTodosByUsers } from '../controllers/admin.controller'
+import { getUsers, listAllTodos, getTodosPerUser, averageTodosByUsers, dailyCompletedTaskCount } from '../controllers/admin.controller'
 import auth, { isAdmin } from '../middlewares/auth'
 
 const server = express();
@@ -9,5 +9,6 @@ server.get('/user', auth, isAdmin, getUsers);
 server.get('/todo', auth, isAdmin, listAllTodos);
 server.get("/todo-per-user", auth, isAdmin, getTodosPerUser);
 server.get("/avg-todo", auth, isAdmin, averageTodosByUsers);
+server.get('/completed/daily-stats', auth, isAdmin, dailyCompletedTaskCount)
 
 export default server;
