+++ src/routes/todoRoutes.ts
@@ -1,7 +1,7 @@
 import express from 'express'
 import auth from '../middlewares/auth'
 import { validateTodoBody } from '../middlewares/bodyValidation'
-import { createTodos, listTodos, updateTodos, deleteTodos, dailyCompletedTaskCount } from '../controllers/todo.controller'
+import { createTodos, listTodos, updateTodos, deleteTodos } from '../controllers/todo.controller'
 import { validateId } from '../middlewares/todoValidations'
 import { findTodoById } from '../middlewares/todoFinder'
 
@@ -12,5 +12,4 @@ server.post('/', auth, validateTodoBody, createTodos);
 server.get('/', auth, listTodos);
 server.patch('/:id', auth, validateId, findTodoById, validateTodoBody, updateTodos);
 server.delete('/:id', auth, validateId, findTodoById, deleteTodos);
-server.get('/completed/daily-stats', auth, dailyCompletedTaskCount)
 export default server;
